ref: move task-card to common components
