+++ frontend/src/app/common-components/task-card/task-card.component.scss
